Also move the existing login components to the new structure
