+++ src/components/structures/login/Login.js
@@ -22,7 +22,7 @@ var sdk = require('matrix-react-sdk');
 var Signup = require("matrix-react-sdk/lib/Signup");
 var PasswordLogin = require("matrix-react-sdk/lib/components/views/login/PasswordLogin");
 var CasLogin = require("matrix-react-sdk/lib/components/views/login/CasLogin");
-var ServerConfig = require("./ServerConfig");
+var ServerConfig = require("../../views/login/ServerConfig");
 
 /**
  * A wire component which glues together login UI components and Signup logic
