+++ src/components/structures/login/Registration.js
@@ -21,8 +21,8 @@ var React = require('react');
 var sdk = require('matrix-react-sdk');
 var MatrixClientPeg = require('matrix-react-sdk/lib/MatrixClientPeg');
 var dis = require('matrix-react-sdk/lib/dispatcher');
-var ServerConfig = require("./ServerConfig");
-var RegistrationForm = require("./RegistrationForm");
+var ServerConfig = require("../../views/login/ServerConfig");
+var RegistrationForm = require("../../views/login/RegistrationForm");
 var CaptchaForm = require("matrix-react-sdk/lib/components/views/login/CaptchaForm");
 var Signup = require("matrix-react-sdk/lib/Signup");
 var MIN_PASSWORD_LENGTH = 6;
